Wait longer to ask for references if the user is typing
Got annoyed with it constantly asking for references as I was typing. Technically, the cursor is changing position! I get it! But it's distracting.

Now, we'll wait a bit longer for typing to end before we start asking.

Also added `skipCurrentReference` (defaults to true) if we should ignore the reference that the cursor is currently inside.